update logout and design
