--- app/src/main/java/se/mau/group12/assigment3/ui/LogoutFragment.java
@@ -1,47 +1,0 @@
-package se.mau.group12.assigment3.ui;
-
-import android.content.Intent;
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import se.mau.group12.assigment3.MainActivity;
-import se.mau.group12.assigment3.R;
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link LogoutFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class LogoutFragment extends Fragment {
-
-    public static LogoutFragment newInstance() {
-        return new LogoutFragment();
-    }
-
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
-                             @Nullable Bundle savedInstanceState) {
-        View view= inflater.inflate(R.layout.fragment_logout, container, false);
-        Intent intent = new Intent(getContext(), MainActivity.class);
-        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK|Intent.FLAG_ACTIVITY_CLEAR_TASK);
-        startActivity(intent);
-        return view;
-    }
-
-
-
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-    }
-
-
-}
