+++ app/src/main/res/layout/recycler_row.xml
@@ -9,7 +9,6 @@
         android:layout_width="411dp"
         android:layout_height="101dp"
         android:layout_marginTop="63dp"
-        android:padding="20dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
@@ -28,8 +27,8 @@
             android:id="@+id/textViewTitle"
             android:layout_width="wrap_content"
             android:layout_height="46dp"
+            android:layout_marginLeft="140dp"
             android:layout_marginTop="20dp"
-            android:layout_marginLeft="180dp"
             android:layout_weight="2"
             android:gravity="center"
             android:text="Squat"
